+++ src/main/java/timeUnits/units/Hour.java
@@ -1,0 +1,11 @@
+package timeUnits.units;
+
+import timeUnits.TimeUnit;
+
+//
+public class Hour extends TimeUnit {
+    //
+    public Hour() {
+        super("Hour", "h");
+    }
+}
