+++ src/main/java/timeUnits/units/Minute.java
@@ -1,0 +1,11 @@
+package timeUnits.units;
+
+import timeUnits.TimeUnit;
+
+//
+public class Minute extends TimeUnit {
+    //
+    public Minute() {
+        super("Minute", "min");
+    }
+}
