+++ src/main/java/timeUnits/units/Second.java
@@ -1,0 +1,11 @@
+package timeUnits.units;
+
+import timeUnits.TimeUnit;
+
+//
+public class Second extends TimeUnit {
+    //
+    public Second() {
+        super("Second", "s");
+    }
+}
